requirements.txt
kafka-python
pytest
requests
pylint
psycopg2-binary